dockerfolder/requirements.txt
fastapi==0.79.1
grpc==1.0.0
grpcio==1.46.3
matplotlib==3.4.3
numpy==1.21.2
pandas==1.4.1
Pillow==9.2.0
requests==2.26.0
scikit_image==0.18.3
skimage==0.0
torch==1.11.0
torchtext==0.12.0
torchvision==0.12.0
tqdm==4.62.3
transformers==4.21.1
uvicorn==0.18.2
